rename rendering files
