+++ Fumen Utils_files/rendering/board-render-fumen.js
